# core-lending-credit-cards-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-credit-cards
  description: Credit Cards Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/credit-cards
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/cc-revolving-lines/{ccRevolvingLineId}:
    get:
      tags:
        - CcRevolvingLine
      summary: Get a revolving line by ID
      operationId: getById
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcRevolvingLineDTO'
    put:
      tags:
        - CcRevolvingLine
      summary: Update a revolving line
      operationId: update
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CcRevolvingLineDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcRevolvingLineDTO'
    delete:
      tags:
        - CcRevolvingLine
      summary: Delete a revolving line
      operationId: delete
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/cc-revolving-lines:
    get:
      tags:
        - CcRevolvingLine
      summary: List or search revolving lines
      operationId: findAll
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.accountId
          in: query
          description: Exact filter for account id
          required: false
          schema:
            type: string
        - name: filters.cardId
          in: query
          description: Exact filter for card id
          required: false
          schema:
            type: string
        - name: filters.productId
          in: query
          description: Exact filter for product id
          required: false
          schema:
            type: string
        - name: filters.creditLimit
          in: query
          description: Exact filter for credit limit
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[creditLimit].from
          in: query
          description: Filter credit limit from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[creditLimit].to
          in: query
          description: Filter credit limit to value
          required: false
          schema:
            type: string
        - name: filters.interestRate
          in: query
          description: Exact filter for interest rate
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[interestRate].from
          in: query
          description: Filter interest rate from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[interestRate].to
          in: query
          description: Filter interest rate to value
          required: false
          schema:
            type: string
        - name: filters.currentBalance
          in: query
          description: Exact filter for current balance
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[currentBalance].from
          in: query
          description: Filter current balance from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[currentBalance].to
          in: query
          description: Filter current balance to value
          required: false
          schema:
            type: string
        - name: filters.availableCredit
          in: query
          description: Exact filter for available credit
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[availableCredit].from
          in: query
          description: Filter available credit from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[availableCredit].to
          in: query
          description: Filter available credit to value
          required: false
          schema:
            type: string
        - name: filters.revolveStatus
          in: query
          description: Exact filter for revolve status
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - CLOSED
              - DEFAULT
              - SUSPENDED
              - OVERLIMIT
        - name: filters.remarks
          in: query
          description: Exact filter for remarks
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCcRevolvingLineDTO'
    post:
      tags:
        - CcRevolvingLine
      summary: Create a new revolving line
      operationId: create
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CcRevolvingLineDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcRevolvingLineDTO'
...
components:
  schemas:
    CcRevolvingLineDTO:
      required:
        - accountId
        - availableCredit
        - cardId
        - creditLimit
        - currentBalance
        - interestRate
        - revolveStatus
      type: object
      properties:
        ccRevolvingLineId:
          type: string
          format: uuid
          readOnly: true
        accountId:
          type: string
          format: uuid
        cardId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        creditLimit:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        interestRate:
          maximum: 1
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        currentBalance:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: -999999999.99
          exclusiveMinimum: false
          type: number
        availableCredit:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        revolveStatus:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - CLOSED
            - DEFAULT
            - SUSPENDED
            - OVERLIMIT
        remarks:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseCcRevolvingLineDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CcRevolvingLineDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...